# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Black Pear Joggers API",
        "contact": {
            "name": "Paul Evans",
            "url": "https://bpj.org.uk",
            "email": "paul3vanz@gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/fetch/athlete/{athleteId}": {
            "get": {
                "tags": [
                    "Fetch"
                ],
                "summary": "Fetch and store one athlete's performances and rankings from Power of 10",
                "operationId": "244f49e72a03cfaae4d86f560bc08117",
                "parameters": [
                    {
                        "name": "athleteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/fetch/performances/{athleteId}": {
            "get": {
                "tags": [
                    "Fetch"
                ],
                "summary": "Fetch and store one athlete's performances from Power of 10",
                "operationId": "8561fc7ad3b57a026c9ecd31ed75657a",
                "parameters": [
                    {
                        "name": "athleteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/fetch/rankings/{athleteId}": {
            "get": {
                "tags": [
                    "Fetch"
                ],
                "summary": "Fetch and store one athlete's runBritain handicap history from Power of 10",
                "operationId": "a9c2467973af7754e785fb67d1819fa0",
                "parameters": [
                    {
                        "name": "athleteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}